fix: two SonarCloud reliability bugs (graphql unsigned negate, redis NaN)
- graphql.cpp: ipart is unsigned (dodges signed-overflow UB while accumulating), so the
  negative-int path did -ipart, a modular unsigned negation (cpp:S876). Negate in signed space.
- redis_resp.cpp: NaN was 0.0/0.0 - identical operands on '/' (cpp:S1764). Produce it as
  infinity*0.0 instead (differing operands, still no <math.h>, matching the file's -inf idiom).
native_graphql + native_redis green.

+++ src/services/graphql/graphql.cpp
@@ -274,7 +274,9 @@ bool parse_value(Lex &L, DetwsGqlValue *v)
         else
         {
             v->type = DetwsGqlType::DETWS_GQL_INT;
-            v->i = neg ? -ipart : ipart;
+            // Negate in signed space: ipart is unsigned (to dodge signed-overflow UB while
+            // accumulating), so -ipart would be a modular unsigned negation, not arithmetic negation.
+            v->i = neg ? -(long long)ipart : (long long)ipart;
         }
         return true;
     }
